worker/requirements.txt
torch==2.6.0
torchvision==0.21.0
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
